fix ToString을 View에 위임
